Docs: More JSDoc.
