+++ docs/examples/zh/geometries/ParametricGeometry.html
@@ -27,7 +27,7 @@
 		<h2>代码示例</h2>
 
 		<code>
-		const geometry = new THREE.ParametricGeometry( THREE.ParametricGeometries.klein, 25, 25 );
+		const geometry = new THREE.ParametricGeometry( klein, 25, 25 );
 		const material = new THREE.MeshBasicMaterial( { color: 0x00ff00 } );
 		const klein = new THREE.Mesh( geometry, material );
 		scene.add( klein );
